requirements.txt
conformer
s3tokenizer
librosa
resemble-perth
huggingface_hub
safetensors
transformers
diffusers
torchaudio
torchvision
einops
runpod>=1.7.7
